[Inductor Cutlass backend] Disabling dynamic EVT fusions
After testing and profiling Cutlass-based EVT fusions, we decided not to move forward with this feature of the Inductor Cutlass backend for the time being, since the small potential performance gains are not worth the considerable increase in Inductor codebase complexity.

This commit disables the EVT fusions and makes sure all tests still pass. In further commits, the corresponding functionality can now be safely removed.

[Inductor Cutlass backend] Removing dynamic EVT fusions

After testing and profiling Cutlass-based EVT fusions, we decided not to move forward with this feature of the Inductor Cutlass backend for the time being, since the small potential performance gains are not worth the considerable increase in Inductor codebase complexity.

This commit removes the code for the dynamic CUTLASS EVT epilogue codegen.

ghstack-source-id: f9911b58664ed2c5e3ece3277a2b3306c69fe9ad
Pull Request resolved: https://github.com/pytorch/pytorch/pull/119010