[aot_inductor] replace TORCH_CHECK with AOTI_CHECK in the generate cpp code
In some cases where we have TORCH_CHECK in loops, it may cause the host
compiler to spend hours optimizing the run_impl function. This PR
mitigated the issue by replacing TORCH_CHECK with a custom AOTI_CHECK,
where we force the underneath assert function to be noinline.

If forcing noinline caused any serious perf regression, we could
either add an option to turn on/off enable noinline. Or, we could
another an option to just turn AOTI_CHECK into a no-op, similar
to the ```assert``` macro from cassert.

ghstack-source-id: f03eb9cd42cd68390c3788955980314d9a5e37fb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/119220