[inductor] Add split scan kernel
this adds a new type of triton kernel in which data is persistent but the
reduction dimension is split over multiple blocks (up to the entire kernel).
though this is called a reduction dimension, in actuality we only support scans.
because of this limitation, i have to be able to block fusions of split scan
operations with reductions so chose to add a new `ir.splitscan` node which
is identical but allows for differentiation in the scheduler.

the split scan kernel is also the first to require an additional workspace buffer
which is used to communicate between cuda blocks. this is slightly tricky as we
the exact scratch space requirement isn't known until the grid size is calculated.
here i workaround the issue by setting a minimum rblock size and always allocating
to the maximum possible grid size for a given input tensor.

ghstack-source-id: fc6bb15ed6d2a0cc17fb402a1b6922df3ed231b9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117992