Support unique id for Tensor Storage Object (#143093)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/143093

PyTorch execution trace records tensor storage data in the trace. The tensor storage data includes storage id, offset, number of elements, and number of byte for each element. PARAM et-replay uses this information to allocate/free the tensors.

However, the current implementation of generating tensor storage id does not guarantee it is unique. ExecutionTraceObserver maintains a lookup table to map the memory address of the tensor storage object to an unique id. If a new memory address is found, it will be put into that hash table and associate it to a new id.

This implementation does not guarantee the storage object is unique since the memory that the address points to may be released and then re-allocated to a different tensor storage object.

To fix it, a static global id is used to generate the unique id for storage object. The global id is an atomic variable and initialized to 0. Then when a new storage object is created, its id is set to the global id, and then bump up the global id.

Test Plan: buck2 run mode/opt caffe2/test:torch -- -r test_torch.TestTorchDeviceTypeCPU.test_storage_*

Differential Revision: D66849516